backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
